requirements.txt
openai
streamlit
llama-index
langchain
python-dotenv
nltk
modal